FINERACT-2080: Extract complex loan actions out from Loan entity
